#110: more refactoring
