--- main_test.go
@@ -1,9 +1,0 @@
-package traefik_oidc_auth
-
-import (
-	"testing"
-)
-
-func Test(t *testing.T) {
-
-}
